requirements.txt
# Core dependencies
click>=8.0.0
numpy>=1.24.0
tqdm>=4.65.0

# StringGen dependencies
# (minimal, mostly stdlib)

# SCAudit dependencies
langchain>=0.1.0
langchain-community>=0.1.0
langchain-openai>=0.1.0
openai>=1.0.0
anthropic>=0.18.0
sqlalchemy>=2.0.0
pysqlcipher3>=1.0.0
pandas>=2.0.0
faiss-cpu>=1.7.0
torch>=2.0.0
transformers>=4.30.0
tenacity>=8.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
pylint>=2.17.0
mypy>=1.0.0

# Optional GPU support
# faiss-gpu>=1.7.0